:recycle: :sparkles: Refactoring  code and implement config
