+++ core/src/main/java/fr/eatroulette/core/plugin/PluginManager.java
@@ -1,4 +1,4 @@
-package fr.eatroulette.core.plugins;
+package fr.eatroulette.core.plugin;
 
 import java.io.File;
 import java.io.IOException;
@@ -22,7 +22,7 @@ public class PluginManager {
 
 
     public PluginManager(){
-        this.pluginFolder = new File("plugin/");
+        this.pluginFolder = new File(PluginManagerConfig.PLUGIN_DIR);
         this.loadAllJar();
     }
 
@@ -73,7 +73,7 @@ public class PluginManager {
         File f = this.hashMap.get(pluginName);
 
         URL url = f.toURI().toURL(); //Because File.toURL is deprecated and the doc tell to do like this
-        System.out.println(url.toString());
+//        System.out.println(url.toString());
         //Init classLoader
         ClassLoader classLoader = new URLClassLoader(new URL[] {url}, getClass().getClassLoader());
 
@@ -85,19 +85,21 @@ public class PluginManager {
             JarEntry entry = enumeration.nextElement();
             if(entry.getName().endsWith(".class")){
                 className = entry.getName().substring(0, entry.getName().length() - 6).replace("/", ".");
+//                System.out.println(className);
+                if(className.endsWith(PluginManagerConfig.PLUGIN_CLASS)){
 
-                if(className.equals("Test")){
                     Class<?> loadedClass = Class.forName(className, true, classLoader);
                     Constructor<?> loadedClassContructor = loadedClass.getConstructor();
 
                     Object instanceOfLodadClass = loadedClassContructor.newInstance();
-                    Method method = loadedClass.getMethod("getName");
+                    Method method = loadedClass.getMethod(PluginManagerConfig.PLUGIN_METHOD);
 
-                    String name = (String) method.invoke(instanceOfLodadClass);
+//                    String name = (String) method.invoke(instanceOfLodadClass);
+                    method.invoke(instanceOfLodadClass);
 
-                    System.out.println("REFLECT NAME : " + name);
+//                    System.out.println("REFLECT NAME : " + name);
                 }
-                System.out.println("Found class => "+entry.getName());
+//                System.out.println("Found class => "+entry.getName());
             }
         }
     }
