+++ core/src/main/java/fr/eatroulette/core/plugin/PluginManagerConfig.java
@@ -1,0 +1,10 @@
+package fr.eatroulette.core.plugin;
+
+public class PluginManagerConfig {
+
+    public static final String PLUGIN_DIR = "plugins/";
+    public static final String PLUGIN_CLASS = "EatRouletteEntry";
+    public static final String PLUGIN_METHOD = "run";
+
+
+}
